Reorganizing packages a bit
