+++ core/src/icbat/games/tradesong/game/workshops/ItemProducer.java
@@ -1,7 +1,6 @@
 package icbat.games.tradesong.game.workshops;
 
 import icbat.games.tradesong.game.Item;
-import icbat.games.tradesong.game.Workshop;
 
 /***/
 public interface ItemProducer extends Workshop {
